services/airflow/requirements.txt
# services/airflow/requirements.txt (For Docker)
yfinance
pandas
numpy
requests
praw
psycopg2-binary